+++ pkg/policy/manager.go
@@ -268,7 +268,7 @@ func (m *ManagerImpl) loadRemotePolicy(ctx context.Context, remotePolicy *Remote
 		}
 	}
 	if policyFilePath == "" {
-		return nil, fmt.Errorf("failed to find policy file; policy.%#v not found in %s", defaultSupportedPolicyExtensions, policyFolder)
+		return nil, fmt.Errorf("failed to find policy file in root directory; expected policy.hcl not found in %s", policyFolder)
 	}
 	m.logger.Debug("policy file found", "path", policyFilePath)
 
@@ -474,6 +474,9 @@ func readPolicy(policyPath, policyFolder string) (*Policies, error) {
 	parser := config.NewParser()
 	policiesRaw, diags := parser.LoadHCLFile(policyPath)
 	if diags != nil && diags.HasErrors() {
+		for _, d := range diags {
+			hclog.Default().Error(d.Error())
+		}
 		return nil, fmt.Errorf("failed to load policy file: %#v", diags.Error())
 	}
 	return decodePolicy(policiesRaw, diags, policyFolder)