+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [v0.16.2] - 2021-11-29
 
 ### :rocket: Added
-* Added support for telemetry, to gain better insight on usage to improve features and tool performance. For additional info see [docs]() [#280](https://github.com/cloudquery/cloudquery/pull/280).
+* Added support for telemetry, to gain better insight on usage to improve features and tool performance. For additional info see [docs](https://docs.cloudquery.io/docs/cli/telemetry) [#280](https://github.com/cloudquery/cloudquery/pull/280).
 * Added support for executing policy in policy [#302](https://github.com/cloudquery/cloudquery/issues/302)
 
 ### :spider: Fixed